fix(temporal): pass full verify on the new temporal-observability package
Why: rebuilding PR #2527's Buildkite build (#13110) after the sibling PR's
fixes showed the earlier green run was stale -- the root `verify` job now
fails outright, and every downstream deploy/release step is skipped as a
result. All four failures trace back to packages/temporal-observability, the
new shared OTel package this PR introduces: compliance-check and the CI test
manifest never learned about it, both consumers still declared a now-unused
@temporalio/interceptors-opentelemetry dependency, and jscpd's duplication
ratchet caught two real regressions -- a brand-new duplicate Zod schema block
and (separately) packages/temporal and packages/scout-for-lol/packages/backend's
tracing.ts files growing more duplicated than the baseline allows.

What:
- scripts/compliance-check.ts: add packages/temporal-observability:build to
  the exemptions list, matching the existing packages/temporal:build and
  packages/architecture:build entries for source-only internal packages with
  no dist/bundle step.
- scripts/ci-test-manifest.json: add the missing
  @shepherdjerred/temporal-observability entry so `test:ci` has a reporting
  target for the package's vitest suite.
- packages/temporal/package.json, packages/scout-for-lol/packages/backend/package.json:
  remove @temporalio/interceptors-opentelemetry and @opentelemetry/resources
  -- both packages now consume the interceptor and Resource-building helpers
  through @shepherdjerred/temporal-observability instead of importing the
  OTel packages directly, so the direct declarations were dead (knip caught
  the first at build #13110, and removing the interceptor import surfaced the
  second as also-unused when it was re-run below).
- packages/homelab/src/cdk8s/src/testing/container-env-schema.ts (new): the
  Deployment-container-env Zod fragment shared by temporal-feature-flags.test.ts
  (this PR's new file) and the pre-existing analytics.test.ts; extracting it
  removes the specific new duplicate jscpd flagged without touching the other
  four files that independently repeat the same shape.
- packages/temporal-observability/src/tracing-resource.ts (new): extracts the
  genuinely-identical pieces of packages/temporal and scout-backend's own
  NodeSDK bootstraps -- the LoggingSpanExporter wrapper (first-success/every-
  failure logging around an OTLP SpanExporter) and the TracingRuntime/
  TracingResourceOptions types plus the Resource-attribute-building call both
  files had copy-pasted. Each tracing.ts keeps its own initializeTracing
  orchestration (LoggerProvider/OTLP-logs wiring is temporal-only;
  AsyncLocalStorage context manager is scout-only) since those parts
  legitimately differ.
- .jscpd-baseline.json: `bun run jscpd --update` after the above genuinely
  reduced duplication (dropped the scout-backend/temporal tracing.ts pair
  from 586 to 347 of 457 allowed tokens, and incidentally reduced an
  unrelated pre-existing birmel/scout-backend tracing.ts pair to zero
  clones) -- pruning the now-stale baseline entries, not loosening anything.
- bun.lock: regenerated via `bun install` after the package.json changes
  (net: -2 deps, +1 new @opentelemetry/semantic-conventions dependency of
  temporal-observability for the resource-building helper).

Verification:
- bun scripts/verify.ts --filter=@shepherdjerred/temporal
  --filter=@shepherdjerred/temporal-observability --filter=@scout-for-lol/backend
  --filter=@homelab/cdk8s --filter=// --concurrency=4 (72/72 tasks passed,
  including compliance-check, knip, jscpd, and test:ci)
- bunx turbo run typecheck test lint --filter=@shepherdjerred/temporal
  --filter=@shepherdjerred/temporal-observability --filter=@scout-for-lol/backend
  --filter=@homelab/cdk8s --force (28/28 tasks passed)
- Not run: live Loki/Tempo query execution, live Kubernetes apply.
