+++ packages/scout-for-lol/packages/backend/package.json
@@ -54,7 +54,6 @@
     "@opentelemetry/context-async-hooks": "^2.7.1",
     "@opentelemetry/core": "^2.7.1",
     "@opentelemetry/exporter-trace-otlp-http": "^0.221.0",
-    "@opentelemetry/resources": "^2.7.1",
     "@opentelemetry/sdk-node": "^0.221.0",
     "@opentelemetry/sdk-trace-base": "^2.7.1",
     "@opentelemetry/semantic-conventions": "^1.40.0",
@@ -73,7 +72,6 @@
     "@temporalio/activity": "1.22.0",
     "@temporalio/client": "1.22.0",
     "@temporalio/common": "1.22.0",
-    "@temporalio/interceptors-opentelemetry": "1.22.0",
     "@temporalio/worker": "1.22.0",
     "@trpc/server": "11",
     "ai": "^7.0.77",
