+++ packages/temporal/package.json
@@ -53,7 +53,6 @@
     "@opentelemetry/core": "^2.7.1",
     "@opentelemetry/exporter-logs-otlp-http": "^0.221.0",
     "@opentelemetry/exporter-trace-otlp-http": "^0.221.0",
-    "@opentelemetry/resources": "^2.7.1",
     "@opentelemetry/sdk-logs": "^0.221.0",
     "@opentelemetry/sdk-node": "^0.221.0",
     "@opentelemetry/sdk-trace-base": "^2.7.1",
@@ -71,7 +70,6 @@
     "@temporalio/activity": "1.22.0",
     "@temporalio/client": "1.22.0",
     "@temporalio/common": "1.22.0",
-    "@temporalio/interceptors-opentelemetry": "1.22.0",
     "@temporalio/worker": "1.22.0",
     "@temporalio/workflow": "1.22.0",
     "@types/sanitize-html": "^2.16.1",
