+++ packages/temporal/src/observability/tracing.ts
@@ -2,14 +2,6 @@ import { NodeSDK } from "@opentelemetry/sdk-node";
 import { OTLPTraceExporter } from "@opentelemetry/exporter-trace-otlp-http";
 import { OTLPLogExporter } from "@opentelemetry/exporter-logs-otlp-http";
 import {
-  resourceFromAttributes,
-  type Resource,
-} from "@opentelemetry/resources";
-import {
-  ATTR_SERVICE_NAME,
-  ATTR_SERVICE_VERSION,
-} from "@opentelemetry/semantic-conventions";
-import {
   diag,
   DiagLogLevel,
   trace,
@@ -22,16 +14,19 @@ import {
 import { logs as logsAPI } from "@opentelemetry/api-logs";
 import {
   BatchSpanProcessor,
-  type ReadableSpan,
-  type SpanExporter,
   type SpanProcessor,
 } from "@opentelemetry/sdk-trace-base";
 import {
   BatchLogRecordProcessor,
   LoggerProvider,
 } from "@opentelemetry/sdk-logs";
-import { ExportResultCode, type ExportResult } from "@opentelemetry/core";
 import { buildArchiveSpanProcessor } from "@shepherdjerred/llm-observability";
+import {
+  buildTracingResource,
+  LoggingSpanExporter,
+  type TracingResourceOptions,
+  type TracingRuntime,
+} from "@shepherdjerred/temporal-observability/tracing-resource";
 import { createStructuredLogger } from "./logging.ts";
 
 const DEFAULT_OTLP_ENDPOINT = "http://tempo.tempo.svc.cluster.local:4318";
@@ -44,18 +39,6 @@ let batchProcessor: BatchSpanProcessor | undefined;
 let loggerProvider: LoggerProvider | undefined;
 let logRecordProcessor: BatchLogRecordProcessor | undefined;
 let tracingRuntime: TracingRuntime | undefined;
-
-export type TracingRuntime = {
-  readonly processor: SpanProcessor;
-  readonly resource: Resource;
-};
-
-export type TracingResourceOptions = {
-  readonly environment?: string;
-  readonly namespace?: string;
-  readonly taskQueue?: string;
-  readonly workerRole?: string;
-};
 
 const jsonLog = createStructuredLogger("observability.tracing");
 
@@ -78,48 +61,6 @@ const diagLogger: DiagLogger = {
   },
 };
 
-class LoggingSpanExporter implements SpanExporter {
-  private readonly inner: SpanExporter;
-  private firstSuccessLogged = false;
-
-  constructor(inner: SpanExporter) {
-    this.inner = inner;
-  }
-
-  export(
-    spans: ReadableSpan[],
-    resultCallback: (result: ExportResult) => void,
-  ): void {
-    this.inner.export(spans, (result) => {
-      if (result.code === ExportResultCode.SUCCESS) {
-        if (!this.firstSuccessLogged) {
-          this.firstSuccessLogged = true;
-          jsonLog("info", "OTLP trace export succeeded (first batch)", {
-            spanCount: spans.length,
-          });
-        }
-      } else {
-        jsonLog("error", "OTLP trace export failed", {
-          spanCount: spans.length,
-          error:
-            result.error instanceof Error
-              ? result.error.message
-              : "unknown export error",
-        });
-      }
-      resultCallback(result);
-    });
-  }
-
-  shutdown(): Promise<void> {
-    return this.inner.shutdown();
-  }
-
-  forceFlush(): Promise<void> {
-    return this.inner.forceFlush?.() ?? Promise.resolve();
-  }
-}
-
 export function initializeTracing(
   options: TracingResourceOptions = {},
 ): TracingRuntime | undefined {
@@ -134,19 +75,13 @@ export function initializeTracing(
   const otlpEndpoint = Bun.env["OTLP_ENDPOINT"] ?? DEFAULT_OTLP_ENDPOINT;
   const serviceName = Bun.env["TELEMETRY_SERVICE_NAME"] ?? DEFAULT_SERVICE_NAME;
   const serviceVersion = Bun.env["GIT_SHA"] ?? Bun.env["VERSION"] ?? "dev";
-  const resource = resourceFromAttributes({
-    [ATTR_SERVICE_NAME]: serviceName,
-    [ATTR_SERVICE_VERSION]: serviceVersion,
-    "deployment.environment.name": options.environment ?? "dev",
-    "temporal.namespace": options.namespace ?? "default",
-    "temporal.task_queue": options.taskQueue ?? "unknown",
-    "temporal.worker.role": options.workerRole ?? "unknown",
-  });
+  const resource = buildTracingResource(serviceName, serviceVersion, options);
 
   const exporter = new LoggingSpanExporter(
     new OTLPTraceExporter({
       url: `${otlpEndpoint}/v1/traces`,
     }),
+    jsonLog,
   );
 
   // Explicit BatchSpanProcessor so we have a handle to forceFlush() before
